# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Tag": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RelationshipGraph": {
				"properties": {
					"id": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"assistant",
							"memory",
							"task"
						]
					},
					"targetId": {
						"type": "string"
					},
					"relationshipType": {
						"type": "string",
						"enum": [
							"related_to",
							"part_of",
							"example_of",
							"derived_from",
							"depends_on",
							"blocks",
							"subtask_of"
						]
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"type",
					"targetId",
					"relationshipType",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MemoryRequest": {
				"properties": {
					"text": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				},
				"required": [
					"text",
					"type"
				],
				"type": "object",
				"description": "-- Data Structures --"
			},
			"TaskResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"output": {},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TaskRequest": {
				"properties": {
					"type": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"type",
					"description"
				],
				"type": "object",
				"additionalProperties": {}
			},
			"RelationshipType": {
				"type": "string",
				"enum": [
					"related_to",
					"part_of",
					"example_of",
					"derived_from",
					"depends_on",
					"blocks",
					"subtask_of"
				]
			},
			"AssistantSuggestion": {
				"properties": {
					"assistantId": {
						"type": "string"
					},
					"score": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"assistantId",
					"score"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AssistantEvaluation": {
				"properties": {
					"assistantId": {
						"type": "string"
					},
					"successRate": {
						"type": "number",
						"format": "double"
					},
					"feedbackAverage": {
						"type": "number",
						"format": "double"
					},
					"tasksCompleted": {
						"type": "number",
						"format": "double"
					},
					"tasksFailed": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"assistantId",
					"successRate",
					"feedbackAverage",
					"tasksCompleted",
					"tasksFailed"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "gpt-assistants-api",
		"version": "0.2.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/tag-extra/entity/{entityId}/{entityType}": {
			"get": {
				"operationId": "GetTagsByEntity",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Tag"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"TagExtra"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "entityType",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"memory",
								"assistant",
								"task"
							]
						}
					}
				]
			}
		},
		"/tag-extra/entity/{entityId}/{entityType}/{tagId}/{isNames}": {
			"post": {
				"operationId": "AddTagToEntity",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"tags": [
					"TagExtra"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "entityType",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"memory",
								"assistant",
								"task"
							]
						}
					},
					{
						"in": "path",
						"name": "tagId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "isNames",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/tag-extra/entity/{entityId}/{entityType}/{tagId}": {
			"delete": {
				"operationId": "RemoveTagFromEntity",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"TagExtra"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "entityType",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"memory",
								"assistant",
								"task"
							]
						}
					},
					{
						"in": "path",
						"name": "tagId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/relationship-graph": {
			"get": {
				"operationId": "GetAllRelationships",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RelationshipGraph"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"RelationshipGraph"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "AddRelationship",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"tags": [
					"RelationshipGraph"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RelationshipGraph"
							}
						}
					}
				}
			}
		},
		"/relationship-graph/source/{targetId}": {
			"get": {
				"operationId": "GetRelationshipsBySource",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RelationshipGraph"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"RelationshipGraph"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "targetId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/relationship-graph/source/{targetId}/type/{relationshipType}": {
			"get": {
				"operationId": "GetRelationshipsBySourceAndType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RelationshipGraph"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"RelationshipGraph"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "targetId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "relationshipType",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"assistant",
								"memory",
								"task"
							]
						}
					}
				]
			}
		},
		"/relationship-graph/{id}": {
			"put": {
				"operationId": "UpdateRelationship",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"RelationshipGraph"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RelationshipGraph"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteRelationship",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"RelationshipGraph"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/orchestrator/remember": {
			"post": {
				"operationId": "Remember",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"tags": [
					"Orchestrator"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"tags": {
										"items": {
											"type": "string"
										},
										"type": "array"
									},
									"memory": {
										"$ref": "#/components/schemas/MemoryRequest"
									},
									"assistantId": {
										"type": "string"
									}
								},
								"required": [
									"memory",
									"assistantId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/orchestrator/forget": {
			"post": {
				"operationId": "Forget",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"tags": [
					"Orchestrator"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"memoryId": {
										"type": "string"
									},
									"assistantId": {
										"type": "string"
									}
								},
								"required": [
									"memoryId",
									"assistantId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/orchestrator/delegate-task": {
			"post": {
				"operationId": "DelegateTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Orchestrator"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"tags": {
										"items": {
											"type": "string"
										},
										"type": "array"
									},
									"task": {
										"$ref": "#/components/schemas/TaskRequest"
									},
									"assistantId": {
										"type": "string"
									}
								},
								"required": [
									"task",
									"assistantId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/orchestrator/connect-assistants": {
			"post": {
				"operationId": "ConnectAssistants",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Orchestrator"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"relation": {
										"$ref": "#/components/schemas/RelationshipType"
									},
									"dependentId": {
										"type": "string"
									},
									"primaryId": {
										"type": "string"
									}
								},
								"required": [
									"relation",
									"dependentId",
									"primaryId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/orchestrator/connect-entities": {
			"post": {
				"operationId": "ConnectEntities",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Orchestrator"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"relation": {
										"$ref": "#/components/schemas/RelationshipType"
									},
									"targetId": {
										"type": "string"
									},
									"targetType": {
										"type": "string",
										"enum": [
											"assistant",
											"memory",
											"task"
										]
									},
									"sourceId": {
										"type": "string"
									},
									"sourceType": {
										"type": "string",
										"enum": [
											"assistant",
											"memory",
											"task"
										]
									}
								},
								"required": [
									"relation",
									"targetId",
									"targetType",
									"sourceId",
									"sourceType"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/orchestrator/query-knowledge": {
			"get": {
				"operationId": "QueryKnowledge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Orchestrator"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "assistantId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tags",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orchestrator/suggest-assistants": {
			"post": {
				"operationId": "SuggestAssistants",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AssistantSuggestion"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Orchestrator"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"tags": {
										"items": {
											"type": "string"
										},
										"type": "array"
									},
									"task": {
										"$ref": "#/components/schemas/TaskRequest"
									}
								},
								"required": [
									"task"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/orchestrator/evaluate-performance/{assistantId}": {
			"get": {
				"operationId": "EvaluatePerformance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AssistantEvaluation"
								}
							}
						}
					}
				},
				"tags": [
					"Orchestrator"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}